+++ packages/botonic-react/src/webchat/hooks/use-component-will-mount.ts
@@ -1,0 +1,5 @@
+import { useMemo } from 'react'
+
+export const useComponentWillMount = func => {
+  useMemo(func, [])
+}
