+++ packages/botonic-react/src/webchat/hooks/use-previous.ts
@@ -1,0 +1,9 @@
+import { useEffect, useRef } from 'react'
+
+export function usePrevious(value) {
+  const ref = useRef()
+  useEffect(() => {
+    ref.current = value
+  })
+  return ref.current
+}
